Remove config.Config from wiki package
As we are ready to get rid of the global config.Config
variable this change does exact it for the files in the
wiki package. It also includes refactoring: addition of
the setup functions, renames and marking functions as
test-helpers.
All ruby sidecar dependent tests are assembled into a single one
as a sub-test. It uses a single instance of the ruby service
created internally. The global rubyServer variable and
initialisation code are removed as they are not unused anymore.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699