maintenance: Inject randomness source into `OptimizeReposRandomly()`
The `OptimizeReposRandomly()` function is currently creating an ad-hoc
source of randomness based on the current Unix timestamp. With this
architecture, the call is hard to test as we cannot make it behave
deterministically.

Refactor the code to instead accept the source of randomness via a
parameter. This slightly semantics given that we never reinitialize the
source anymore as we previously did when `OptimizeReposRandomly` was
called. This is perfectly fine though: we do not care whether we
reinitialize or not -- it should be random enough either way.