maintenance: Convert repo optimizer to use new random walker
Now that we have a separate implementation of the random filesystem walk
via the `randomWalker`, this commit converts the `OptimizeRepository`
implementation to use it. This allows us to convert the current
recursive implementation into a simple loop, which is easier to reason
about and also easier to extend in subsequent commits.

No functional changes are expected from this commit.