blob: Implement filter transformer
We're about to introduce a new way of retrieving all LFS pointer blobs
in a repository, no matter whether reachable or not, in order to speed
up retrieval of LFS blobs. Instead of needing to do a complete graph
walk in order to determine all newly pushed objects and thus potential
pointers, we just iterate over pushed objects directly. This will be
done with `git cat-file --batch-all-objects`. But given that it does not
support the same filters which git-rev-list(1), we'll have to do the
filtering ourselves.

This commit introduces the filter in form of an implementation of the
`text/transform.Transformer` interface. Given object information in the
format of `%(objecttype) %(objectsize) %(objectname)`, the transformer
drops every line that is either not a blob or is bigger than 200 bytes.
All the other lines which are potential LFS pointer candidates get
transformed to only print their object names.

This allows us to keep the pipeline from git-cat-file to git-cat-file
short such that we're able to reuse the existing implementation of
`readLFSPointers()` by just transforming the output of `git-cat-file
--batch-all-objects --batch-check="%(objecttype) %(objectsize)
%(objectname)"` and passing the resulting transformed reader into
`readLFSPointers()`.