blob: Implement ListLFSPointers() RPC
We currently have three different functions to retrieve LFS pointers:
one which checks object IDs directly, one which retrieves all reachable
LFS pointers and a third one which does a limited graph walk. In short,
all three of these search LFS pointers by iterating a set of revisions.
The current interface is thus unnecessarily complex given that it has
three limited ways to do the same thing instead of providing one general
implementation which allows both our own API to be more concise as well
as allowing users of the API to be more flexible.

This commit thus implements a replacement interface `ListLFSPointers`.
Instead of restricting users, we simply accept a set of revisions which
we traverse in order to find LFS pointers. Because `GetNewLFSPointers`
allowed users to restrict the graph walk via a set of negative refs, we
also accept the pseudo-revisions `--not` and `--all`.

With this new and simple interface, we can replace all existing usecases
and thus mark the three other RPCs as deprecated.