Pass backchannel id through hooks to HookService
The peer id in the context can be used to identify the backchannel
that is connected to the Praefect that sent the mutator. That Praefect
should receive any votes related to the RPC. However, once the
HookService receives the the request from the hook, the peer id no longer
points to the Praefect that originally sent the request but to the hook
that sent the request. This commit passes the backchannel's id through
the hooks to the HookService so it gets the knowledge which backchannel it
should use to send the votes to the correct Praefect.