blob: Implement ListAllLFSPointers() RPC
In order to verify whether a push is allowed or not, we do a call to
Rails' `/allowed` endpoint. This endpoint does multiple checks: next to
determining whether the reference updates are allowed in the first
place, there's also several checks which inspect all new objects which
are part of the push.

One of these checks is the check for LFS pointers. For each push, we get
a call to `GetNewLFSPointers` which computes the set of new objects and
then extracts all new LFS pointers from this set. If any of the new LFS
pointers does not have a corresponding LFS object in the repository,
then we refuse the push.

Computation of new objects can be heavily expensive though, depending on
the repository's size: we need to do a complete graph walk to correctly
determine preexisting objects and new objects. For big repositories with
lots of references and commits, this can take several seconds and in the
most extreme cases lead to context cancellations as the walks exceed the
30 seconds allowed for those checks. The user cannot do anything about
this, except restricting repository size (which we definitely don't want
to recommend) or disabling LFS pointer checks altogether (potentially
compromising repository consistency).

There is one realization to be had though: when doing pushes into git,
git will first accept all objects into a quarantine environment. As
such, there is a single place which contains all new objects which have
been part of the push. So if we'd be able to just single out pushed
objects and check these instead of doing a graph walk, then we'd start
to scale with push size, not with repository size.

There is an easy way to do this via `git cat-file --batch-all-objects`,
which prints out all of the ODB's objects no matter whether reachable or
not. Given that git spawns processes with the main object directory set
to the quarantine environment and the normal object directory part of
the alternative object directories, the only thing we need to do to
single out only pushed objects is to unset the alternative object
directories: `env --unset=GIT_ALTERNATIVE_OBJECT_DIRECTORIES git
cat-file --batch-all-objects`.

A quick benchmark with gitlab-org/gitlab.git shows that this is much
faster. The following tests have been done by pushing into the target
repository which had the LFS pointer checks as pre-receive hook. Output
has been formatted such that it becomes more readable.

    # 1000 commits with one change each
    $ git push origin master
    Benchmark #1: LFS pointers via rev-list
      Time (mean ± σ):     554.3 ms ±  20.6 ms    [User: 527.5 ms, System: 27.0 ms]
      Range (min … max):   521.9 ms … 590.5 ms    10 runs

    Benchmark #2: LFS pointers via --batch--all-objects
      Time (mean ± σ):       3.8 ms ±   1.6 ms    [User: 5.8 ms, System: 2.5 ms]
      Range (min … max):     2.4 ms …  23.0 ms    555 runs

    Summary
      'LFS pointers via --batch--all-objects' ran
      145.14 ± 59.30 times faster than 'LFS pointers via rev-list'

    # push 100 branches, where each has the same 1000 commits plus one that is different per branch
    $ git push origin $(seq -f 'branch-%g'  100)
    Benchmark #1: LFS pointers via rev-list
      Time (mean ± σ):     620.9 ms ±   7.0 ms    [User: 584.8 ms, System: 36.0 ms]
      Range (min … max):   613.3 ms … 633.1 ms    10 runs

    Benchmark #2: LFS pointers via --batch--all-objects
      Time (mean ± σ):       4.4 ms ±   1.6 ms    [User: 6.3 ms, System: 3.1 ms]
      Range (min … max):     0.2 ms …  26.5 ms    636 runs

    Summary
      'LFS pointers via --batch--all-objects' ran
      140.34 ± 49.49 times faster than 'LFS pointers via rev-list'

    # push of unrelated history to emulate lots of objects (pushing Gitaly into the GitLab repo)
    $ git push origin gitaly/master:refs/heads/gitaly
    Benchmark #1: LFS pointers via rev-list
      Time (mean ± σ):     625.5 ms ±  10.1 ms    [User: 590.0 ms, System: 35.5 ms]
      Range (min … max):   615.3 ms … 651.2 ms    10 runs

    Benchmark #2: LFS pointers via --batch--all-objects
      Time (mean ± σ):       6.4 ms ±   1.5 ms    [User: 7.9 ms, System: 3.8 ms]
      Range (min … max):     2.2 ms …  14.8 ms    467 runs

    Summary
      'LFS pointers via --batch--all-objects' ran
       98.11 ± 23.32 times faster than 'LFS pointers via rev-list'

So even for biggish pushes, `--batch-all-objects` is about 50x faster
than doing the graph walk.

In order to allow Rails to make use of this new way of doing things,
this commit implements a new interface `ListAllLFSPointers()`. In
contrast to the existing-but-deprecated `GetAllLFSPointers()` RPC, it
will return all LFS pointers regardless of their reachability. In order
to only make use of quarantined objects, the caller will then have to
modify the `Repository` message to unset alternative oject directories.