git: Move injection of `pack.windowMemory` configuration
Based on whether a command may generate packfiles or not, we inject the
`pack.windowMemory` configuration into the command's options. This is
currently handled in `combineOpts()`, which is a weird place to handle
this: it's about user-provided command options and thus shouldn't care
about command-specific configuration.

Move assembly of command-specific configuration into `combineArgs()` to
fix this weirdness.