Serve RefTransactionService on the backchannel connection
This commit implements Praefect's backchannel server factory. The
server factory is a function that returns a new server that should
serve on the backchannel connection Gitaly establishes to Praefect.
Gitaly only needs to contact Praefect's RefTransactionService, so it
is the only registered service. The interceptor stack in the backchannel
server matches mostly what Praefect generally has giving us observability
into the backchannel server. There few omissions, which are not required
by the backchannel server:

1. Auth interceptor is not added. It's not necessary as only the Gitalys
   Praefect dials can connect to the backchannel server.

2. MethodTypeUnaryInterceptor is not needed as it's only counting the types
   of proxied RPCs.

3. Keep alives are not configured as Praefect and Gitaly are already sending
   keep alives on the Praefect to Gitaly stream. Those keep alives are
   sufficient to keep the underlying network connection open.