move internals of client into an internal package
Praefect requires a more flexible API to plug in its multiplexing
implementation to when dialing to Gitaly. This implementation should
not be part of the public API interface so we don't have to maintain
backwards compatibility. This commit moves the internals of the dialing
into an internal subpackage. This way we can retain the exisiting public
APIs we offer in `client` package and provide more flexible APIs for
internal use in `ìnternal/gitaly/client`.