git: Implement `SupportsConfigEnv()`
With git v2.31.0, a new mechanism to inject configuration entries via
the environment has been implemented which we now want to start make use
of. Given that our minimum git version is still at v2.29.0, we need to
gate usage of this new mechanism behind a feature check.

Implement a new `SupportsConfigEnv()` function which checks whether the
feature is supported or not.