move backchannel's ServerFactory into a struct
This commit shuffles some of the backchannel's client handshake code
around to minimize the dependency drilling later required. Initializing
ClientHandshaker with the logger once allows us to later only inject a
minimal set of dependencies into the dialing functions as we don't have
to drill down the logger which would be needed by the ServerFactory's
ClientHandshake method currently. After this change, the ClientHandshaker
already contains the logger so we don't need to pass the logger down
separately. No functional changes in this commit, only to improve code
ergonomics.