git: Convert option and arg handling functions to receivers
The functions to handle both options and arguments for an
about-to-be-spawned command both are plain non-receiver functions. But
in the following commit, we're about to start using internal state of
the command factory. As a preparatory step, this commit thus converts
both functions to be receiver functions of the `ExecCommandFactory`.