git: Move invalid commands metric into into command factory
The metric which counts the number of invalid commands is currently
hosted as a global metric. Historically, we didn't have any other choice
given that there was no central structure which could host it instead.
But this has changed with the introduction of the git command factory:
in production code, it is constructed once and then injected into all of
its users. It has thus become the obvious place to host this metric.

Move the metric into the command factory to reduce our reliance on
global variables.