Add CheckHostIP=no to SSH auth options for mirroring
With CheckHostIP=yes (the default), the SSH client will proactively
convert a hostname to an IP, and try to add that IP to the known_hosts
file if it isn't already present. Since our UserKnownHostsFile is read-
only, that fails with a harmless (but confusing) warning message.

Adding CheckHostIP=no removed that warning message. It also means that
we no longer try to protect against DNS spoofing, but since ssh will
automatically (try to) add the fingerprint of unknown IPs to the
known_hosts file at present, the value of that right now is effectively
nil.