handle muxed connections in resolvePraefectAddress
With muxed connections, authInfo is never nil. If no transport security
is needed, it is set to the 'insecure' transport credentials.
resolvePraefectAddress was not taking the transport security type into
account and failed when Gitaly was contacted over an insecure muxed TCP
connection. This commit fixes the problem by handling the insecure muxed
TCP connection case as it does without the multiplexing.