objectpool: Use localrepo.Remote interface in `FetchFromOrigin()`
The `FetchFromOrigin()` function is currently using an ad-hoc
implementation to check for the existence of remotes and add/update
them.

Refactor the code to instead use the `localrepo.Remote` interface and
its functions. Note that we cannot use it for the actual fetch though:
we pass an explicit refspec that's different from the configured one,
and that's not yet supported by the interface.