remote: Use atomic fetch in `FetchInternalRemote()`
Transactional behaviour for fetches had to be disabled at some point
because the reference-transaction hook was executed once per updated
reference. Naturally, this can be extremely expensive in cases where
thousands of references are updated, leading to thousands of votes.

Since then, we've upstreamed a new `--atomic` flag into git, which has
been released with git v2.31.0. With this flag, all references are
updated in a single transaction in an all-or-nothing fashion. Which
effectively also means that the reference-transaction hook is now
executed once, only.

Enable this behaviour in `FetchInternalRemote()` if the current git
version supports it. The side effect of making the fetch now an
all-or-nothing operation shouldn't cause any issues for us here given
that we use a forced refspec anyway.