git: Refactor assembly of command config
The code to handle command options and then combines the command's
arguments is a bit hard to read. Most importantly, it's awkward that
`handleOpts()` gets as input a pointer to the command config which it is
about to populate.

Refactor the code to instead have `handleOpts()` return the command
config, highlighting that it is its responsibility to generate the
config. Given that it's not modified by `combineArgs()`, it is converted
to accept the structure as a value instead of as a pointer.