git: Refactor version interface
The version interface is currently quite awkward to use. The user first
obtains the current git version via `git.Version()` and passes that on
to `VersonLessThan()`, which only accepts to stings. Because the former
function spawns a process and the latter function needs to parse both
versions, both functions need error handling. The code thus becomes
needlessly verbose.

Refactor the code to make the `Version` structure public and return it
from the renamed `git.CurrentVersion()` function. This allows to avoid
the second parse and is a nicer interface to work with in general. It
also makes it a lot easier at a later point to implement various
`IsSupported()` functions for various features and thus serves as a
preparation for the following commit.