gitaly-git2go: Rebase merged branch is fastforward
When rebasing a branch that's merged into upstream, all commits already
exist in that branch. That means, there are no more commits to replay on
top of upstream and thus rebase can simply return the HEAD commit of the
upstream branch as the resulting commit for the rebase operation. The
caller can then fast-forward the branch to that commit.