gitaly-git2go: Rebase skips commits upstream has
When a branch was partially merged into upstream, rebase should only
pick commits that are not in upstream.

This change uses merge base between the branch and upstream to determine
which commits do not yet exist in upstream. Only the descendant commits
from merge base are taken into account for the rebase.