Remove incorrect pre-allocation from GetConsistentStorages
GetConsistentStorages previously returned only the consistent
secondaries. Back then, we were removing the primary from the result
set and thus pre-allocated the map to the number of secondaries
ancitipating all secondaries to be consistent.

This does not make sense anymore. We don't remove the primary from
the result set so the initial capacity was wrong, likely resulting
in an extra allocation anyway. With variable replication factor, we
can't even guess the result count from the storages, so let's just
remove the incorrect map pre-allocation.

This commit also renames the `consistentSecondaries` map to
`consistentStorages` for accuracy. `storages` method is removed as it
is no longer used.