git: Update locations which need update on version bumps
We have recently adopted a list of all locations which need to be
updated when the minimum required git version is bumped. Because of the
efforts to distribute a single git distribution across GitLab
distributions with Gitaly's `make git` target, this list is now out of
date:

    - Build images do not need to be updated when bumping the minimum
      required version. They must've been updated beforehand already.

    - CNG has recently started to use `make git` and thus does not need
      updating anymore.

Drop both of these locations.