remote: Fix UpdateRemoteMirror having transactional semantics
While UpdateRemoteMirror performs a mutating action on a remote
repository which potentially is not under our control, it will never
modify the local target repository. It thus doesn't make a lot of sense
to treat it as a mutator: it's never mutating local state, and if it
targets a remote repository which is hosted by Gitaly, too, then it'd
use normal routing and thus have mutator semantics on that remote.

Relabel the RPC as an accessor to stop it from having transactional
semantics: we never create any transactions as nothing changes, which
causes us to always schedule replication jobs. And second, doing the
same push from all three nodes at once is definitely not going to work.