ref: Fix missing votes for `DeleteRefs()` RPC
Production systems currently report that 20% of all RPCs which do not
perform any votes is the `DeleteRefs()` RPC. A first guess is that this
may happen in case a user requests deletions where there are no refs to
be deleted, and that guess would be correct.

But in fact, none of the RPC calls will ever cast any votes: because of
99f0c28c9 (hook: Fix voting on pushes which delete packed references,
2021-02-15), we're forced to filter out any reference-transactions which
consist solely of force deletions. And that's what `DeleteRefs()` does:
it's force-deleting all refs which are requested for deletion, and thus
the reference-transaction hook filters out the votes.

Fix the issue by doing a manual vote on all refs which are about to be
deleted.