transaction: Introduce Vote type
When voting, we currently only accept a byte slice representing the vote
itself. This requires the caller to know how to compute the vote and
thus leaks implementation details across all voters.

Improve the situation by introducing a new `Vote` type, which simply
maps to a fixed-size byte array, along with constructors to create a
`Vote`.