transaction: Introduce new VoteHash type
In order to construct a voting hash from input, the developer can
currently either use `VoteFromData()`, which accepts a plain byte slice,
or alternatively instantiate a SHA1 hasher. The former isn't really
suited to process bigger amounts of data, and the latter is leaking
implementation details of the voting hash to callers.

To improve the situation, this commit introduces a new `VoteHash`
structure which simply wraps a SHA1 hasher.