updateref: Verify sending commands succeeds
There's two writes in `updateref.Updater` to the child process which
aren't checked. In the worst case, this could mean that those writes do
not arrive and leave us in the believe that they are effective, while
indeed they aren't.

Fix this by checking errors of both callsites. While at it, convert both
cases to use `Write()` directly, given that there are no formatting
directives involved.