updateref: Provide function to send "prepare" command
When using git-update-ref(1) with the `--stdin` flag, the caller can
directly control transactional behaviour if he desires to do so. This
includes the verbs "start", "prepare", "commit" and "abort". We
explicitly use "start" and "commit" already to enable transactional
behaviour and to commit changed refs. "abort" is implicitly used when
the command gets killed without a "commit". The "prepare" is also
implicit in the "commit": if the user didn't manually prepare the
transaction, then git will automatically prepare the reference update
for us by locking all refs part of the transaction.

Add an explicit interface for the user such that "prepare" can be called
explicitly: this allows the developer to split up locking of the refs
and committing them such that it's possible to do intermediate steps
between those two stages. One usecase for this is to do manual voting:
all refs to be updated would be queued, the transaction prepared, then a
manual vote happens and only if the vote is successful will the update
be committed.

Currently, this is kind of a best-effort implementation: we do not
verify that commands queued to the updater actually take effect because
we do not check e.g. whether there's a "prepare: ok" line written by
git-update-ref(1). But if the prepare fails, we'd fail at a later point
anyway when committing. So while this is an annoyance and may cause the
transactional vote to succeed even if locking refs failed, this should
not be a case of data inconsistencies. Regardless of that, this issue
should be fixed in a future iteration.