repository: Fix fetching in-memory remotes with SSH params
When fetching remotes, the caller can pass the SSH key which the caller
wants to use for the given fetch. To inject this into git, Gitaly will
set up the GIT_SSH_COMMAND environment variable and pass it to the
git-fetch(1) invocation. But while we set up this variable for named
remotes, we do not for in-memory ones.

There is no real reason to not do this: if given an in-memory SSH
remote, then it's perfectly expected that we continue to use the given
private SSH key. So this commit fixes the issue by just setting up the
SSH command for both in-memory and named remotes.

Unfortunately, we currently do not have nice infrastructure for SSH
fetches which would allow us to properly test this. To verify it works
as expected, the testcase thus simply uses a wrapper for the git
executable and inspects that it gets passed GIT_SSH_COMMAND with correct
arguments.