README.md: Document updated Go version requirements
Our README.md still pretends that we support Go 1.13 and newer. This
has changed a few releases ago already and thus isn't accurate: first,
we never suppord "and newer", but only the latest two releases. And
second, we now support Go 1.15 and 1.16.

Update the file to explicitly spell out the two supported releases.