tool-version: Introduce support for asdf
GitLab has recently started to support asdf as part of the developer's
workflow to track and install required tools in the correct version.
Many of GitLab's projects have already started to provide a
`.tool-version` file, which tells asdf which versions of which tool are
required in order to build and test a given project.

Introduce a `.tool-version` file for Gitaly, too. For now, it only
tracks the currently supported Ruby and Golang versions.