Generate NOTICE using GOOS=linux
`make notice` produces a different NOTICE file when ran on macOS and
Linux. This is due to some linux build tagged files including Linux
specific dependencies, namely the cgroup packages. This commit fixes
the problem by running the NOTICE generation with GOOS=linux so the
resulting NOTICE file matches regardless whether it was generated on
macOS or Linux.