conflicts: Fix argument tests always missing a user
The tests which verify that `ResolveConflicts()` correctly raises an
`InvalidArgument` when passed various invalid arguments didn't check for
the returned error message. Of course, all tests were succeeding because
of something that we didn't actually want to test: none of the requests
have the `User` field set up.

Tighten tests by asserting the received error message and add the `User`
argument everywhere. This reveals that we actually do not check for an
unset repository, which this commit also fixes.