conflicts: Do not accept non-OIDs in the Go port of `ResolveConflicts()`
When passing non-OID revisions as our or their commit OID, then the old
Ruby implementation of  `ResolveConflicts()` used to return an error
saying that the OID could not be parsed correctly. In Go, we happily
accept valid revisions and resolve them to their referenced OID.

Fix the bug by verifying that passed OIDs are indeed valid OIDs and add
a test which verifies we raise the same error message as Ruby does.