conflicts: Raise error message if resolution is the same as contents
The `ResolveConflicts()` RPC accepts as the resolution of a merge
conflict either a list of linewise resolutions or the complete contents
of the merged file. In the latter case, the Ruby port verifies that
resolved contents and conflicting contents are indeed different such
that we do not commit a yet-unresolved merge conflict. This behaviour
got lost during the Go port: we just commit the contents without any
safety check.

Fix the issue by comparing given contents and the merged file's
contents, raising an error if they're the same and add a test which
verifies that both Ruby and Go port behave the same.